% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_dmd.R
\name{alloc_dmd}
\alias{alloc_dmd}
\title{Allocates a constrained demand between receiving entities}
\usage{
alloc_dmd(dataset, DFU, Period, Demand, Opening, Supply)
}
\arguments{
\item{dataset}{a dataframe with the demand and supply features for an item per period}

\item{DFU}{name of an item, a SKU, or a node like an item x location}

\item{Period}{a period of time monthly or weekly buckets for example}

\item{Demand}{the quantity of an item planned to be consumed in units for a given period}

\item{Opening}{the opening inventories of an item in units at the beginning of the horizon}

\item{Supply}{the quantity of an item planned to be supplied in units for a given period}
}
\value{
a dataframe with the calculated Projected Inventories and Coverages as well as the total Constrained Demand and the allocated demand between receiving entities
}
\description{
Allocates a constrained demand between receiving entities
}
\examples{
alloc_dmd(dataset = alloc_data, DFU, Period, Demand, Opening, Supply)

}
