% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blueprint}
\alias{blueprint}
\title{blueprint}
\format{
A data frame with 520 rows and 7 variables
}
\usage{
data(blueprint)
}
\description{
This dataset contains the basic features to calculate projected inventories and coverages and also 2 additional info: a minimum and maximum targets of stock coverage.
We can apply on it the proj_inv() function, it will return calculated projected inventories and coverages as well as an analysis of the position of the projected inventories versus the minimum and maximum stocks targets.
}
\details{
\itemize{
\item DFU, an item
\item Period, a date
\item Demand, a consumption in units
\item Opening.Inventories, available inventories at the beginning in units
\item Supply.Plan, a Replenishment Plan in units
\item Min.Stocks.Coverage, a Minimum Stocks Targets in number of Periods
\item Max.Stocks.Coverage, a Maximum Stocks Targets in number of Periods

}
}
\author{
Nicolas Nguyen \email{nikonguyen@yahoo.fr}
}
