% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blueprint_drp}
\alias{blueprint_drp}
\title{blueprint_drp}
\format{
A data frame with 520 rows and 9 variables
}
\usage{
data(blueprint_drp)
}
\description{
This dataset contains the basic features to calculate a Replenishment Plan (also called DRP) and its related projected inventories and coverages.
We can apply on it the drp() function, it will return the calculated Replenishment Plan and its related projected inventories and coverages.
}
\details{
\itemize{
\item DFU, an item
\item Period, a date
\item Demand, a consumption in units
\item Opening, available inventories at the beginning in units
\item Supply, a Replenishment Plan in units
\item FH, defines the Frozen and Free Horizon. It has 2 values: Frozen or Free. If Frozen : no calculation of Replenishment Plan yet, the calculation starts when the period is defined as Free. We can use this parameter to consider some defined productions plans or supplies (allocations, workorders,...) in the short-term for example.
\item SSCov, the Safety Stock Coverage, expressed in number of periods
\item DRPCovDur the Frequency of Supply, expressed in number of periods
\item MOQ the Multiple Order Quantity, expressed in units, 1 by default or a Minimum Order Quantity

}
}
\author{
Nicolas Nguyen \email{nikonguyen@yahoo.fr}
}
