% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blueprint_light}
\alias{blueprint_light}
\title{blueprint_light}
\format{
A data frame with 520 rows and 5 variables
}
\usage{
data(blueprint_light)
}
\description{
This dataset contains the basic features to calculate projected inventories and coverages.
Just 5 features are needed for this: a DFU, a Period, a Demand, an initial Opening Inventory and a Supply Plan.
We can apply on it the light_proj_inv() function, it will return calculated projected inventories and coverages.
}
\details{
\itemize{
\item DFU, an item
\item Period, a date
\item Demand, a consumption in units
\item Opening, available inventories at the beginning in units
\item Supply, a Replenishment Plan in units

}
}
\author{
Nicolas Nguyen \email{nikonguyen@yahoo.fr}
}
