% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-clone_plan.R
\name{clone_plan}
\alias{clone_plan}
\title{Calculates actual duration of tasks accounting for holidays}
\usage{
clone_plan(projectname)
}
\arguments{
\item{projectname}{project name}
}
\description{
Produces an actual planning calendar with public holidays
}
