% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_vol.R
\name{cov_vol}
\alias{cov_vol}
\title{Calculates the calculated coverage expressed in volume of a coverage initially expressed in periods of time}
\usage{
cov_vol(dataset, DFU, Period, Demand, Coverage)
}
\arguments{
\item{dataset}{a dataframe with the demand and supply features for an item per period}

\item{DFU}{name of an item, a SKU, or a node like an item x location}

\item{Period}{a period of time monthly or weekly buckets for example}

\item{Demand}{the quantity of an item planned to be consumed in units for a given period}

\item{Coverage}{the defined Coverage, expressed in number of periods}
}
\value{
a dataframe with the calculated coverage expressed in volume
}
\description{
Calculates the calculated coverage expressed in volume of a coverage initially expressed in periods of time
}
\examples{
cov_vol(dataset = cov_vol_data, DFU, Period, Demand, Coverage)

}
