% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cov_vol_data}
\alias{cov_vol_data}
\title{cov_vol_data}
\format{
A data frame with 520 rows and 4 variables
}
\usage{
data(cov_vol_data)
}
\description{
This dataset contains the basic features to convert a coverage expressed in periods into a coverage expressed in units
We can apply on it the cov_vol() function, it will return the related coverage in units
}
\details{
\itemize{
\item DFU, an item
\item Period, a date
\item Demand, a consumption in units
\item Coverage the targeted coverage that we want to convert, expressed in number of periods

}
}
\author{
Nicolas Nguyen \email{nikonguyen@yahoo.fr}
}
