% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-create-calendar.R
\name{create_calendar}
\alias{create_calendar}
\title{Creates a calendar with public holidays}
\usage{
create_calendar(holiday_dates = NULL, weekend_days = c("saturday", "sunday"))
}
\arguments{
\item{holiday_dates}{vector of dates of public holidays}

\item{weekend_days}{string of dates of public holidays}
}
\value{
Planning calendar
}
\description{
Produces an actual planning calendar with public holidays
}
\examples{
#To include Christmas
create_calendar(holiday_dates = as.Date("2023-12-25"),  weekend_days = c("saturday", "sunday"))
}
