% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-create_plan.R
\name{create_plan}
\alias{create_plan}
\title{Calculates actual duration of tasks accounting for holidays}
\usage{
create_plan(projectname)
}
\arguments{
\item{projectname}{project name}
}
\description{
Produces an actual planning calendar with public holidays
}
