% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_monthly_dmd}
\alias{demo_monthly_dmd}
\title{demo_monthly_dmd}
\format{
A data frame with 24 rows and 3 variables
}
\usage{
data(demo_monthly_dmd)
}
\description{
This dataset contains a set of Monthly Demand for two Products.
There are 3 variables: a DFU, a Monthly Period, a Monthly Demand.
The idea is to use this dataset to convert the Demand from Monthly into Weekly bucket.
We can apply on this dataset the month_to_week() function, it will create a weekly bucket Period and convert the Demand from Monthly into Weekly bucket.
}
\details{
\itemize{
\item DFU, an item
\item Period, a date in monthly format
\item Demand, a consumption in units

}
}
\author{
Nicolas Nguyen \email{nikonguyen@yahoo.fr}
}
