% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_to_cov.R
\name{inv_to_cov}
\alias{inv_to_cov}
\title{Calculates the Projected Coverages related to Projected Inventories}
\usage{
inv_to_cov(dataset, DFU, Period, Inventories, Demand)
}
\arguments{
\item{dataset}{a dataframe with the demand forecasts and projected inventories for an item per period}

\item{DFU}{name of an item, a SKU, or a node like an item x location}

\item{Period}{a period of time monthly or weekly buckets for example}

\item{Inventories}{the projected inventories that we aim to convert into their related projected coverages}

\item{Demand}{the quantity of an item planned to be consumed in units for a given period}
}
\value{
a dataframe with the calculated Projected Coverages
}
\description{
Calculates the Projected Coverages related to Projected Inventories
}
\examples{
inv_to_cov(dataset = inventories_data, DFU, Period, Inventories, Demand)

}
