% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventories_data}
\alias{inventories_data}
\title{inventories_data}
\format{
A data frame with 72 rows and 4 variables
}
\usage{
data(inventories_data)
}
\description{
This dataset contains the basic features to convert projected inventories into their related projected coverage expressed in periods
We can apply on it the inv_to_cov() function, it will return the related coverage in periods
}
\details{
\itemize{
\item DFU, an item
\item Period, a date
\item Inventories, projected inventories in units that we want to convert in coverage
\item Demand, a consumption in units

}
}
\author{
Nicolas Nguyen \email{nikonguyen@yahoo.fr}
}
