% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-plot_plan.R
\name{plot_plan}
\alias{plot_plan}
\title{Calculates actual duration of tasks accounting for holidays}
\usage{
plot_plan(
  plan,
  plan_start_date,
  holiday_dates = NULL,
  weekend_days = c("saturday", "sunday"),
  filename
)
}
\arguments{
\item{plan}{vector of dates of public holidays}

\item{weekend_days}{string of dates of public holidays}
}
\value{
Actual duration of tasks
}
\description{
Produces an actual planning calendar with public holidays
}
\examples{
#To include Christmas
create_calendar(plan, holiday_dates = as.Date("2023-12-25"),  weekend_days = c("saturday", "sunday"))
}
