% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slob}
\alias{slob}
\title{slob}
\format{
A data frame with 44 rows and 4 variables
}
\usage{
data(slob)
}
\description{
This dataset contains the detailed Opening Inventories for two Products.
There are 4 variables: a DFU, a Period, a Demand and the breakdown of the Opening Inventories by expiry date or minimum Remaining Shelf Life for use.
The idea is to use this dataset to calculate the Short Shelf Life quantities, called here SSL Qty.
We can apply on this dataset the ssl() function, it will calculate a SSL Qty field.
}
\details{
\itemize{
\item DFU, an item
\item Period, a date in monthly format
\item Demand, a consumption in units
\item Opening, the breakdown of the opening inventories in units by expiry date

}
}
\author{
Nicolas Nguyen \email{nikonguyen@yahoo.fr}
}
