% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_to_week.R
\name{month_to_week}
\alias{month_to_week}
\title{Calculates the Projected Inventories and Coverages as well as the Constrained Demand and informs a Tag about the part of the Demand already covered by the Opening Inventories}
\usage{
month_to_week(dataset, DFU, Period, Demand)
}
\arguments{
\item{dataset}{a dataframe with the demand in monthly bucket for each item}

\item{DFU}{name of an item, a SKU, or a node like an item x location}

\item{Period}{a monthly period of time that we want to convert into weekly buckets}

\item{Demand}{the quantity of an item planned to be consumed in units for a given period}
}
\value{
a dataframe with the Demand expressed in weekly buckets for each item
}
\description{
Calculates the Projected Inventories and Coverages as well as the Constrained Demand and informs a Tag about the part of the Demand already covered by the Opening Inventories
}
\examples{
month_to_week(dataset = demo_monthly_dmd, DFU, Period, Demand)

}
